requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==1.0.0
anyio==4.6.2.post1
APScheduler==3.10.4
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
binaryornot==0.4.4
cattrs==23.2.3
certifi==2024.6.2
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
distlib==0.3.8
Django==5.0.7
django-agent-trust==1.1.0
django-formtools==2.5.1
django-otp==1.5.1
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
docutils==0.21.2
esbonio==0.16.4
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
lsprotocol==2023.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pre-commit==3.8.0
pybit==5.8.0
pycryptodome==3.21.0
pygls==1.3.1
Pygments==2.18.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.6
requests==2.32.3
rich==13.7.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spotipy==2.24.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
virtualenv==20.26.3
websocket-client==1.8.0
yarl==1.9.4
